Moved code for proxy to service.
